# docs/openapi/operator-mgmt.yaml
---
openapi: 3.0.3
info:
  title: java-operator API
  version: 1.2.1
servers:
- url: https://localhost:8443
  description: Auto generated value
- url: https://0.0.0.0:8443
  description: Auto generated value
paths:
  /apis:
    get:
      tags:
      - Discovery Api
      responses:
        "200":
          description: OK
  /apis/operator.enonic.cloud/v1:
    get:
      tags:
      - Info Api
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
  /apis/operator.enonic.cloud/v1/mutations:
    post:
      tags:
      - Mutation Api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdmissionReview'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionReview'
  /apis/operator.enonic.cloud/v1/operator/version:
    get:
      tags:
      - Operator Api
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorVersion'
  /apis/operator.enonic.cloud/v1/validations:
    post:
      tags:
      - Admission Api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdmissionReview'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionReview'
  /apis/operator.enonic.cloud/v1/xp7/{namespace}/{name}/{nodegroup}/mgmt/content/projects/list:
    get:
      tags:
      - Xp 7 Management Api
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: nodegroup
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Xp7MgmtProject'
  /apis/operator.enonic.cloud/v1/xp7/{namespace}/{name}/{nodegroup}/mgmt/idproviders/list:
    get:
      tags:
      - Xp 7 Management Api
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: nodegroup
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Xp7MgmtIdProvider'
  /apis/operator.enonic.cloud/v1/xp7/{namespace}/{name}/{nodegroup}/mgmt/repo/snapshots/list:
    get:
      tags:
      - Xp 7 Management Api
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: nodegroup
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Xp7MgmtSnapshotsList'
  /apis/operator.enonic.cloud/v1/xp7/{namespace}/{name}/{nodegroup}/mgmt/webapps/list:
    get:
      tags:
      - Xp 7 Management Api
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: nodegroup
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Xp7MgmtWebapp'
components:
  schemas:
    AdmissionRequest:
      type: object
      properties:
        dryRun:
          type: boolean
        kind:
          $ref: '#/components/schemas/GroupVersionKind'
        name:
          type: string
        namespace:
          type: string
        object:
          $ref: '#/components/schemas/KubernetesResource'
        oldObject:
          $ref: '#/components/schemas/KubernetesResource'
        operation:
          type: string
        options:
          $ref: '#/components/schemas/KubernetesResource'
        requestKind:
          $ref: '#/components/schemas/GroupVersionKind'
        requestResource:
          $ref: '#/components/schemas/GroupVersionResource'
        requestSubResource:
          type: string
        resource:
          $ref: '#/components/schemas/GroupVersionResource'
        subResource:
          type: string
        uid:
          type: string
        userInfo:
          $ref: '#/components/schemas/UserInfo'
    AdmissionResponse:
      type: object
      properties:
        allowed:
          type: boolean
        auditAnnotations:
          type: object
          additionalProperties:
            type: string
        patch:
          type: string
        patchType:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        uid:
          type: string
        warnings:
          type: array
          items:
            type: string
    AdmissionReview:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        request:
          $ref: '#/components/schemas/AdmissionRequest'
        response:
          $ref: '#/components/schemas/AdmissionResponse'
    GroupVersionKind:
      type: object
      properties:
        kind:
          type: string
        group:
          type: string
        version:
          type: string
    GroupVersionResource:
      type: object
      properties:
        group:
          type: string
        resource:
          type: string
        version:
          type: string
    KubernetesResource:
      type: object
    ListMeta:
      type: object
      properties:
        continue:
          type: string
        remainingItemCount:
          format: int64
          type: integer
        resourceVersion:
          type: string
        selfLink:
          type: string
    OperatorVersion:
      type: object
      properties:
        gitCommit:
          type: string
        gitTags:
          type: string
        buildDate:
          type: string
        version:
          type: string
        gitTreeState:
          type: string
    Status:
      type: object
      properties:
        apiVersion:
          type: string
        kind:
          type: string
        metadata:
          $ref: '#/components/schemas/ListMeta'
        code:
          format: int32
          type: integer
        details:
          $ref: '#/components/schemas/StatusDetails'
        message:
          type: string
        reason:
          type: string
        status:
          type: string
    StatusCause:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        reason:
          type: string
    StatusDetails:
      type: object
      properties:
        kind:
          type: string
        causes:
          type: array
          items:
            $ref: '#/components/schemas/StatusCause'
        group:
          type: string
        name:
          type: string
        retryAfterSeconds:
          format: int32
          type: integer
        uid:
          type: string
    UserInfo:
      type: object
      properties:
        extra:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        groups:
          type: array
          items:
            type: string
        uid:
          type: string
        username:
          type: string
    Xp7MgmtBranch:
      type: object
      properties:
        name:
          type: string
        sites:
          type: array
          items:
            $ref: '#/components/schemas/Xp7MgmtSite'
    Xp7MgmtIdProvider:
      type: object
      properties:
        key:
          type: string
        displayName:
          type: string
        description:
          type: string
    Xp7MgmtProject:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        parent:
          type: string
        branches:
          type: array
          items:
            $ref: '#/components/schemas/Xp7MgmtBranch'
    Xp7MgmtSite:
      type: object
      properties:
        displayName:
          type: string
        path:
          type: string
        language:
          type: string
    Xp7MgmtSnapshotsList:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Xp7MgmtSnapshotsListSnapshot'
    Xp7MgmtSnapshotsListSnapshot:
      type: object
      properties:
        name:
          type: string
        reason:
          type: string
        state:
          type: string
        timestamp:
          type: string
        indices:
          type: array
          items:
            type: string
    Xp7MgmtWebapp:
      type: object
      properties:
        name:
          type: string